+++ nauticradio/CHANGES.md
@@ -1,4 +1,9 @@
 ## Changes
 
+**Version 0.2.2**
+- removed wget dependency.
+- added direct download of files.
+- added compatibility with parsers main script
+
 **Version 0.2.1**
 - initial version. Based on my c't Uplink parser.